requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
Cython==0.29.34
decorator==4.4.2
docutils==0.20.1
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
Kivy==2.2.1
Kivy-Garden==0.1.5
kivymd==1.1.1
moviepy==1.0.3
numpy==1.25.2
Pillow==10.0.0
ping3==4.0.4
proglog==0.1.10
Pygments==2.15.1
pyjnius==1.5.0
python-settings==0.2.2
requests==2.31.0
tqdm==4.66.1
uri==2.0.1
urllib3==2.0.3
